Reverted remaining int32-related changes, and updated the util Makefile so that a 32-bit architecture is enforced at build-time.

+++ include/qtext.h
@@ -12,7 +12,7 @@
 struct qtmsg {
 	int	msgnum;
 	char	delivery;
-	uint32_t  offset,
+	long    offset,
 		  size;
 };
